+++ stix2validator/enums.py
