+++ stix2validator/output.py
