+++ stix2validator/test/attack_pattern_tests.py
