+++ stix2validator/test/custom_obj_tests.py
