+++ stix2validator/test/identity_tests.py
