+++ stix2validator/test/indicator_tests.py
