+++ stix2validator/test/malware_tests.py
