+++ stix2validator/test/marking_definition_tests.py
