+++ stix2validator/test/relationship_tests.py
