+++ stix2validator/test/threat_actor_tests.py
