+++ stix2validator/test/tool_tests.py
