+++ stix2validator/test/vulnerability_tests.py
