+++ tox.ini
@@ -3,7 +3,7 @@ envlist = py26, py27, py33, py34, py35
 
 [testenv]
 commands =
-    nosetests validator
+    nosetests stix2validator
 deps=
     nose
     tox
